Remove subgraph_detector from inference/analysis to the common framework/ir directory. test=develop
