+++ paddle/fluid/inference/analysis/ir_passes/CMakeLists.txt
@@ -1,13 +1,10 @@
-cc_library(subgraph_detector SRCS subgraph_detector.cc subgraph_util.cc DEPS proto_desc)
-if(WITH_TESTING)
-  add_dependencies(subgraph_detector gtest)
-endif()
+cc_library(subgraph_util SRCS subgraph_util.cc DEPS subgraph_detector)
 
 if (WITH_GPU AND TENSORRT_FOUND)
-  cc_library(tensorrt_subgraph_pass SRCS tensorrt_subgraph_pass.cc DEPS subgraph_detector tensorrt_op_teller)
+  cc_library(tensorrt_subgraph_pass SRCS tensorrt_subgraph_pass.cc DEPS subgraph_util tensorrt_op_teller)
 
   set(analysis_deps ${analysis_deps}
-          subgraph_detector tensorrt_subgraph_pass
+          subgraph_util tensorrt_subgraph_pass
           CACHE INTERNAL "")
 
   set(pass_file ${PADDLE_BINARY_DIR}/paddle/fluid/inference/api/paddle_inference_pass.h)
@@ -16,10 +13,10 @@ if (WITH_GPU AND TENSORRT_FOUND)
 endif()
 
 if (ANAKIN_SUBGRAPH) 
-  cc_library(anakin_subgraph_pass SRCS anakin_subgraph_pass.cc DEPS subgraph_detector anakin_op_teller)
+  cc_library(anakin_subgraph_pass SRCS anakin_subgraph_pass.cc DEPS subgraph_util anakin_op_teller)
 
   set(analysis_deps ${analysis_deps}
-          subgraph_detector anakin_subgraph_pass
+          subgraph_util anakin_subgraph_pass
           CACHE INTERNAL "")
 
   set(pass_file ${PADDLE_BINARY_DIR}/paddle/fluid/inference/api/paddle_inference_pass.h)
